rename common/contract-details to contract-seo
