+++ web/pages/[username]/[contractSlug].tsx
@@ -30,7 +30,7 @@ import { CPMMBinaryContract } from 'common/contract'
 import { AlertBox } from 'web/components/widgets/alert-box'
 import { useTracking } from 'web/hooks/use-tracking'
 import { useSaveReferral } from 'web/hooks/use-save-referral'
-import { getContractOGProps, getSeoDescription } from 'common/contract-details'
+import { getContractOGProps, getSeoDescription } from 'common/contract-seo'
 import { ContractDescription } from 'web/components/contract/contract-description'
 import { ContractLeaderboard } from 'web/components/contract/contract-leaderboard'
 import { ContractsGrid } from 'web/components/contract/contracts-grid'
